+++ database/migrations/2020_11_27_060228_create_thoi_gian_chieus_table.php
@@ -20,7 +20,7 @@ class CreateThoiGianChieusTable extends Migration
             //
             $table->bigIncrements('MaThoiGianChieu');
             $table->time('ThoiGianChieu');
-            $table->integer('TrangThai')->default(0);
+            $table->tinyInteger('TrangThai')->default(0);
             $table->timestamp('ThoiGianTao')->useCurrent();
             $table->timestamp('ThoiGianCapNhatCuoi')->useCurrent();
         });
